pyproject.toml
[tool.poetry]
name = "izanami"
version = "0.1.0"
description = ""
authors = ["boke0 <speken00.tt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.13"
mitama = "^4.2.9"
docker = "^4.4.4"
invoke = "^1.5.0"

[tool.poetry.dev-dependencies]
unittest = "^0.0"
flake8 = "^3.8.4"
isort = "^5.7.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
